results/beyond_wpt/localhost/localhost_cookies.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/cookies/resources/cookie-helper.sub.js"></script>
<script>
function set_cookie_via_dom_test(options) {
  promise_test(async (t) => {
    var name = options.prefix + "prefixtestcookie";
    await test_driver.delete_all_cookies()
    // erase_cookie_from_js(name, options.params);
    t.add_cleanup(test_driver.delete_all_cookies);
    var value = "" + Math.random();
    document.cookie = name + "=" + value + ";" + options.params;

    assert_dom_cookie(name, value, options.shouldExistInDOM);
  }, options.title);
}

  set_cookie_via_dom_test({
    prefix: "__Host-",
    params: "Secure; Path=/;",
    shouldExistInDOM: true,
    title: "__Host: Non-secure localhost: 'Secure; Path=/'"
  });

  set_cookie_via_dom_test({
    prefix: "__Secure-",
    params: "Secure; Path=/;",
    shouldExistInDOM: true,
    title: "__Secure: Non-secure localhost:'Secure; Path=/;'"
  });

  set_cookie_via_dom_test({
    prefix: "",
    params: "Secure; Path=/",
    shouldExistInDOM: true,
    title: "Secure attribute: Non-secure localhost: 'Path=/;"
  });
</script>
